Move a few splatter classes into API
